XXXXXX

+++ app/controllers/carts_controller.rb
@@ -4,8 +4,8 @@ class CartsController < ApplicationController
     @ticket_ids.each do |ticket_id|
       @purchase = Purchase.new(ticket_id: ticket_id, user_id: session[:user_id])
       unless @purchase.save
-        Purchase.where(ticket_id: @ticket_ids).delete_all
-        return redirect_to cart_path(session[:user_id]), success: "#{ticket_id}で失敗"
+        Purchase.where(ticket_id: @ticket_ids, user_id: session[:user_id]).delete_all
+        return redirect_to cart_path(session[:user_id]), danger: "#{ticket_id}は既に売り切れです"
       end
     end
     Cart.where(user_id: session[:user_id]).delete_all
